Re-org ‘Colors’
Apparently altering the theme provider causes all the Emotion snaps to need updating
