+++ src/global_styling/variables/colors.ts
@@ -6,20 +6,10 @@
  * Side Public License, v 1.
  */
 
-import { saturate, shade, tint } from '../../services/color';
-import { computed } from '../../services/theme/utils';
 import {
   ColorModeSwitch,
   StrictColorModeSwitch,
 } from '../../services/theme/types';
-import {
-  makeDisabledContrastColor,
-  makeHighContrastColor,
-} from '../../services/color/contrast';
-
-/*
- * TYPES
- */
 
 /**
  * Top 5 colors
@@ -159,132 +149,11 @@ export type _EuiThemeConstantColors = {
   ink: string;
 };
 
-export type _EuiThemeColors = _EuiThemeBrandColors &
+export type _EuiThemeColorsMode = _EuiThemeBrandColors &
   _EuiThemeBrandTextColors &
   _EuiThemeShadeColors &
   _EuiThemeSpecialColors &
   _EuiThemeTextColors;
 
-/*
- * LIGHT THEME
- * Only split up in the light theme to access the keys by section in the docs
- */
-
-export const brand_colors: _EuiThemeBrandColors = {
-  primary: '#006BB4',
-  accent: '#DD0A73',
-  success: '#017D73',
-  warning: '#F5A700',
-  danger: '#BD271E',
-};
-
-export const brand_text_colors: _EuiThemeBrandTextColors = {
-  primaryText: computed(makeHighContrastColor('colors.primary')),
-  accentText: computed(makeHighContrastColor('colors.accent')),
-  successText: computed(makeHighContrastColor('colors.success')),
-  warningText: computed(makeHighContrastColor('colors.warning')),
-  dangerText: computed(makeHighContrastColor('colors.danger')),
-};
-
-export const shade_colors: _EuiThemeShadeColors = {
-  emptyShade: '#FFF',
-  lightestShade: '#F5F7FA',
-  lightShade: '#D3DAE6',
-  mediumShade: '#98A2B3',
-  darkShade: '#69707D',
-  darkestShade: '#343741',
-  fullShade: '#000',
-};
-
-export const special_colors: _EuiThemeSpecialColors = {
-  body: computed(([lightestShade]) => tint(lightestShade, 0.5), [
-    'colors.lightestShade',
-  ]),
-  highlight: '#FFFCDD',
-  disabled: computed(([darkestShade]) => tint(darkestShade, 0.7), [
-    'colors.darkestShade',
-  ]),
-  disabledText: computed(makeDisabledContrastColor('colors.disabled')),
-  shadow: computed(({ colors }) =>
-    shade(saturate(colors.mediumShade, 0.25), 0.5)
-  ),
-};
-
-export const text_colors: _EuiThemeTextColors = {
-  text: computed(makeHighContrastColor('colors.darkestShade')),
-  title: computed(
-    ([{ text, body }]) => makeHighContrastColor(shade(text, 0.5))(body),
-    ['colors']
-  ),
-  subdued: computed(makeHighContrastColor('colors.mediumShade')),
-  link: computed(([primaryText]) => primaryText, ['colors.primaryText']),
-};
-
-export const light_colors: _EuiThemeColors = {
-  ...brand_colors,
-  ...shade_colors,
-  ...special_colors,
-  // Need to come after special colors so they can react to `body`
-  ...brand_text_colors,
-  ...text_colors,
-};
-
-/*
- * DARK THEME
- */
-
-export const dark_shades: _EuiThemeShadeColors = {
-  emptyShade: '#1D1E24',
-  lightestShade: '#25262E',
-  lightShade: '#343741',
-  mediumShade: '#535966',
-  darkShade: '#98A2B3',
-  darkestShade: '#D4DAE5',
-  fullShade: '#FFF',
-};
-
-export const dark_colors: _EuiThemeColors = {
-  // Brand
-  primary: '#1BA9F5',
-  accent: '#F990C0',
-  success: '#7DE2D1',
-  warning: '#FFCE7A',
-  danger: '#F66',
-  ...dark_shades,
-
-  // Special
-  body: computed(([lightestShade]) => shade(lightestShade, 0.45), [
-    'colors.lightestShade',
-  ]),
-  highlight: '#2E2D25',
-  disabled: computed(([darkestShade]) => tint(darkestShade, 0.7), [
-    'colors.darkestShade',
-  ]),
-  disabledText: computed(makeDisabledContrastColor('colors.disabled')),
-  shadow: computed(({ colors }) =>
-    shade(saturate(colors.mediumShade, 0.25), 0.5)
-  ),
-
-  // Need to come after special colors so they can react to `body`
-  ...brand_text_colors,
-
-  // Text
-  text: '#DFE5EF',
-  title: computed(([text]) => text, ['colors.text']),
-  subdued: computed(makeHighContrastColor('colors.mediumShade')),
-  link: computed(([primaryText]) => primaryText, ['colors.primaryText']),
-};
-
-/*
- * FULL
- */
-
-export type EuiThemeColors = StrictColorModeSwitch<_EuiThemeColors> &
+export type _EuiThemeColors = StrictColorModeSwitch<_EuiThemeColorsMode> &
   _EuiThemeConstantColors;
-
-export const colors: EuiThemeColors = {
-  ghost: '#FFF',
-  ink: '#000',
-  LIGHT: light_colors,
-  DARK: dark_colors,
-};
